+++ fs/types/effects/module.f.ts
@@ -5,7 +5,7 @@
  */
 
 export type Operation =
-    readonly[string, (_: never) => unknown]
+    readonly[string, (..._: readonly never[]) => unknown]
 
 export type Effect<O extends Operation, T> = {
     value: Value<O, T>
@@ -22,7 +22,7 @@ export type DoKPR<O extends Operation, T, K extends string, PR extends readonly[
     readonly[K, PR[0], (_: PR[1]) => Effect<O, T>]
 
 export type Pr<O extends Operation, K extends O[0]> =
-    O extends readonly[K, (_: infer P) => infer R] ? readonly[P, R] : never
+    O extends readonly[K, (...args: infer P) => infer R] ? readonly[P, R] : never
 
 export type DoK<O extends Operation, T, K extends O[0]> =
     DoKPR<O, T, K, Pr<O, K>>
@@ -51,25 +51,18 @@ export type Return<O extends Operation> = F<O>[1]
 
 export const do_ =
     <O extends Operation>(cmd: O[0]) =>
-    (param: Param<O>): Effect<O, Return<O>> =>
-    doFull(cmd, param, pure)
-
-export const doRest =
-    <O extends Operation>(cmd: O[0]) =>
     (...param: Param<O>): Effect<O, Return<O>> =>
-    do_(cmd)(param as Param<O>)
+    doFull(cmd, param as Param<O>, pure)
 
 export const begin: Effect<never, void> = pure(undefined)
 
 export type ToAsyncOperationMap<O extends Operation> = {
-    readonly [K in O[0]]: (payload: Pr<O, K>[0]) => Promise<Pr<O, K>[1]>
+    readonly [K in O[0]]: (...payload: Pr<O, K>[0]) => Promise<Pr<O, K>[1]>
 }
 
 export type F<O extends Operation> = Pr<O, O[0]>
 
-export type Func<O extends Operation> = (_: Param<O>) => Effect<O, Return<O>>
-
-export type RestFunc<O extends Operation> = (..._: Param<O>) => Effect<O, Return<O>>
+export type Func<O extends Operation> = (..._: Param<O>) => Effect<O, Return<O>>
 
 export type ListEffect<O extends Operation, T> =
     Effect<O, readonly[T, ListEffect<O, T>] | undefined>